Fix issue with Flight Launch clusters
Clusters launched by Flight Launch include a hash of the launcher's
email address in their stack name.  That hash is not included in the
cluster name.  Previously, tracon assumed that the domain name and
cluster name would correctly identify the AWS stack.

In order to work with this `alces compute ...` needs sends the
"qualified cluster name".  For Flight Launch clusters this is the
cluster name plus the email hash.

That change is sufficient to allow tracon to correctly obtain the AWS
stack for the cluster.  However, tracon needs to use the "unqualified
cluster name" when adding a queue.  This commit has tracon read the
unqualified cluster name from the `ClusterName` parameter for the
cluster's master stack.